refact: move health tests to module
